Fixed a couple of bugs handling tokens in ParseEngine
Now keywords are properly found and returned as Tokens. hello.m8r still fails because we're not handling multi-char operators yet

+++ lib/m8rscript/src/ParseEngine.cpp
@@ -64,7 +64,7 @@ ParseEngine::Token ParseEngine::getToken()
             return Token::Identifier;
         }
     
-        return static_cast<Token>(uint16_t(result[-1]) + 0x100);
+        return static_cast<Token>(uint16_t(uint8_t(result[-1])) + 0x100);
     }
 
     // FIXME: Add implementaton for multi-char operators
